+++ hw12_13_14_15_calendar/internal/server/http/server.go
@@ -7,12 +7,11 @@ import (
 	"time"
 
 	"github.com/murashko91/otus-go/hw12_13_14_15_calendar/internal/app"
-	"github.com/murashko91/otus-go/hw12_13_14_15_calendar/internal/storage"
 )
 
 type Server struct {
 	server *http.Server
-	app    Application
+	app    app.Application
 	logger app.Logger
 }
 
@@ -21,20 +20,7 @@ type ServerConf struct {
 	Port int
 }
 
-type Application interface {
-	CreateUser(context.Context, storage.User) (storage.User, error)
-	GetUser(context.Context) (storage.User, error)
-	UpdateUser(context.Context, storage.User) (storage.User, error)
-	DeleteUser(context.Context) error
-	CreateEvent(context.Context, storage.Event) (storage.Event, error)
-	UpdateEvent(context.Context, storage.Event) (storage.Event, error)
-	DeleteEvent(context.Context, int) error
-	GetDailyEvents(context.Context, time.Time) ([]storage.Event, error)
-	GetWeeklyEvents(context.Context, time.Time) ([]storage.Event, error)
-	GetMonthlyEvents(context.Context, time.Time) ([]storage.Event, error)
-}
-
-func NewServer(logger app.Logger, app Application, conf ServerConf) *Server {
+func NewServer(logger app.Logger, app app.Application, conf ServerConf) *Server {
 	calendarRouter := http.NewServeMux()
 
 	calendarRouter.Handle("/hello", loggingMiddleware(http.HandlerFunc(helloHandler), logger))
